buy subscription is now ok! we need to refactor wallet

+++ OOD_Project_Backend/Finance/Business/Contracts/IFinanceFacade.cs
@@ -5,5 +5,5 @@ namespace OOD_Project_Backend.Finance.Business.Contracts;
 public interface IFinanceFacade
 {
     Task CreateWallet(UserEntity user,int initBalance = 100);
-    Task<bool> Buy(int userId, int amount, Dictionary<int, double> incomeShareByUserId);
+    Task<bool> Buy(int userId, double amount, Dictionary<int, double> incomeShareByUserId);
 }
